+++ packages/ui-core/src/components/Grid/Grid.tsx
@@ -23,7 +23,7 @@ interface IProps {
 const cn = cnCreate('mfui-grid');
 class Grid extends React.Component<IProps, {}> {
     static propTypes = {
-        hAlign: PropTypes.oneOf(['left', 'center', 'between', 'around']),
+        hAlign: PropTypes.oneOf(['right', 'center', 'between', 'around']),
         vAlign: PropTypes.oneOf(['top', 'bottom', 'center', 'baseline']),
         guttersLeft: PropTypes.oneOf(['large', 'medium']),
         guttersBottom: PropTypes.oneOf(['large', 'medium']),