+++ packages/ui-core/CHANGELOG.md
@@ -4,8 +4,10 @@
 - **Grid**
     - add examples in .mdx
     - fixed value for prop 'hAlign'
+    - added tests
 - **GridColumn**
     - fixed types for props
+    - added tests
 
 ## Release 1.0.5
 - **GridColumn**